refactor: Move MainActivity menu to ProjectFragment
