+++ app/src/main/res/layout/sort_project_dialog.xml
@@ -15,7 +15,7 @@
             android:layout_width="wrap_content"
             android:layout_height="match_parent"
             android:layout_weight="1"
-            android:text="Sort by Name" />
+            android:text="Sort by Project Name" />
 
         <RadioButton
             android:id="@+id/sortByID"
@@ -36,13 +36,13 @@
             android:layout_width="wrap_content"
             android:layout_height="match_parent"
             android:layout_weight="1"
-            android:text="Ascending" />
+            android:text="Ascending (A-Z)" />
 
         <RadioButton
             android:id="@+id/sortOrderDesc"
             android:layout_width="wrap_content"
             android:layout_height="match_parent"
             android:layout_weight="1"
-            android:text="Descending" />
+            android:text="Descending (Z-A)" />
     </RadioGroup>
 </LinearLayout>
