--- app/src/main/res/menu/main_menu.xml
@@ -1,9 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<menu xmlns:android="http://schemas.android.com/apk/res/android">
-
-    <item
-        android:id="@+id/sortProject"
-        android:icon="@drawable/sorting_options_48"
-        android:showAsAction="always"
-        android:title="Sort" />
-</menu>
