+++ app/src/main/res/values/strings.xml
@@ -2650,4 +2650,5 @@ Please check the URL again."</string>
     <string name="text_to_search">Text to search</string>
     <string name="myprojects_list_menu_title_restore_projects">Restore projects</string>
     <string name="search_projects">Filter by name, package, project id</string>
+    <string name="sort_projects">Sort projects</string>
 </resources>
